Make pawn hash an int so we can use array indexes wider than short

+++ EubosChess/src/main/java/eubos/position/IPositionAccessors.java
@@ -12,7 +12,7 @@ public interface IPositionAccessors {
 	public int getMoveNumber();
 	public int getPlyNumber();
 	public long getHash();
-	public short getPawnHash();
+	public int getPawnHash();
 	public String getFen();
 	public boolean isThreefoldRepetitionPossible();
 	public boolean isInsufficientMaterial();
